requirements.txt
asgiref==3.11.0
Django==5.2.8
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.10
python-dotenv==1.0.0
gunicorn==21.2.0
djangorestframework==3.14.0